Fix backend choice collision
Previously it was possible to select both cairo and the web
backend by passing --features=web without having the target be
wasm32.

This attempts to simplfy how we choose the backend.

We will use cairo if it is selected explicitly, or if we're not on
wasm32, windows, or don't see the 'web' feature;

We will use d2d if we're on windows and don't see the 'cairo' feature;

And we will use the web backend if our target arch is wasm32 or
we see the "web" feature.

We will fail to compile if we cannot select a backend.